fastapi/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
Pillow
